+++ CHANGELOG.md
@@ -16,6 +16,8 @@ the detailed section referring to by linking pull requests or issues.
 
 #### Added
 
+- Added login polling to prevent auto-logout when page is open.
+
 #### Changed
 
 #### Removed